Fix compatibility issue
This fixes the compatibility issue with both Windows Explorer and Apple Finder. Like 7zip, we always write out the CRCs in the substream information and never in the folder info.

This work well with solid, consolidate and mixed files.

I also made our reading logic more robust, since up until now, it always expected the CRC for the block to be in the folder info, which isn't even the case for files created by 7zip. We now properly also check the substream info, if there is a single CRC, which we could use. Multiple CRC would hint at multiple files in a solid compressed block, for which each CRC would be only for the compressed data of one of the files.

I also updated the version we write out to 4, since this is this format we are actually writing out.

I also made the compress example a bit more useful for us and added a compatibility test for windows with both solid and non-solid files.

Tested this changes with 7zip, Keka, Finder and Explorer.