Fix compatibility issue
Both the explorer and finder integration can't open the current files. This PR properly implements the behavior that 7zip itself is showing.

We don't store the CRC in the folder section of the unpack info, but instead in the substreams info.

With these changes the archives can properly read with 7zip and also with the explorer and finder integration.